[TIMOB-18078] iOS: Ti.Stream.read() doesn't work with Ti.Filesystem.FileStream

+++ iphone/Classes/TiNetworkSocketTCPProxy.m
@@ -837,7 +837,7 @@ TYPESAFE_SETTER(setError, error, KrollCallback)
             case TO_STREAM: {
                 // Perform the write to stream
                 id<TiStreamInternal> stream = [info valueForKey:@"destination"];
-                int size = [TiUtils intValue:[info valueForKey:@"chunkSize"]];
+                NSInteger size = [TiUtils intValue:[info valueForKey:@"chunkSize"]];
                 KrollCallback* callback = [info valueForKey:@"callback"];
                 
                 TiBuffer* tempBuffer = [[[TiBuffer alloc] _initWithPageContext:[self executionContext]] autorelease];
@@ -854,7 +854,7 @@ TYPESAFE_SETTER(setError, error, KrollCallback)
             case TO_CALLBACK: {
                 // Perform the pump to callback
                 KrollCallback* callback = [info valueForKey:@"callback"];
-                int size = [TiUtils intValue:[info valueForKey:@"chunkSize"]];
+                NSInteger size = [TiUtils intValue:[info valueForKey:@"chunkSize"]];
                 
                 TiBuffer* tempBuffer = [[[TiBuffer alloc] _initWithPageContext:[self executionContext]] autorelease];
                 [tempBuffer setData:[NSMutableData dataWithData:data]];
